requirements.txt
biopython==1.81
Flask==3.0.0
gunicorn==20.1.0